ADD pad transform and split into image and tabular preprocessing
