+++ autoPyTorch/pipeline/components/preprocessing/tabular_preprocessing/scaling/Normalizer.py
@@ -10,7 +10,7 @@ import numpy as np
 from sklearn.compose import make_column_transformer
 from sklearn.preprocessing import Normalizer as SklearnNormalizer
 
-from autoPyTorch.pipeline.components.preprocessing.scaling.base_scaler import BaseScaler
+from autoPyTorch.pipeline.components.preprocessing.tabular_preprocessing.scaling.base_scaler import BaseScaler
 
 
 class Normalizer(BaseScaler):
