+++ autoPyTorch/pipeline/components/preprocessing/tabular_preprocessing/scaling/StandardScaler.py
@@ -7,7 +7,7 @@ from scipy.sparse import issparse
 from sklearn.compose import make_column_transformer
 from sklearn.preprocessing import StandardScaler as SklearnStandardScaler
 
-from autoPyTorch.pipeline.components.preprocessing.scaling.base_scaler import BaseScaler
+from autoPyTorch.pipeline.components.preprocessing.tabular_preprocessing.scaling.base_scaler import BaseScaler
 
 
 class StandardScaler(BaseScaler):
